% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{n_words}
\alias{n_words}
\title{Counting functions}
\usage{
n_words(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Plethora of function to extract features of a text. Re-exported from \code{textfeatures}. Please run \code{?textfeatures::count_functions} for more details.
}
